changed location name to logs
