--- models/locations.js
@@ -1,13 +1,0 @@
-const mongoose = require("mongoose");
-
-const locationSchema = new mongoose.Schema({
-    name: String,
-    location: String,
-    cost: Number,
-    image: [String],
-    willRecommend: Boolean,
-});
-
-const Logs = mongoose.model("Log", locationSchema);
-
-module.exports = Locations;
